Move login into user service
