+++ app/graphql/types/mutation_type.rb
@@ -1,7 +1,7 @@
 module Types
   class MutationType < Types::BaseObject
     field :signup, mutation: Mutations::Users::Signup
-    field :login, mutation: Mutations::Sessions::Login
+    field :login, mutation: Mutations::Users::Login
     field :update_user, mutation: Mutations::Users::Update
     field :delete_user, mutation: Mutations::Users::Delete
   end
