+++ app/middleware/services/users_service.rb
@@ -1,5 +1,5 @@
-module Users
-  class Service < BaseService
+module Services
+  class UsersService < BaseService
 
     def create
       user = User.new params
@@ -7,7 +7,7 @@ module Users
         token = Jwt::TokenGenerator.issue_token user.for_token
         { token: token, errors: [] }
       else
-        { token: 'Invalid', errors: user.errors.full_messages }
+        { token: nil, errors: user.errors.full_messages }
       end
     end
 
@@ -24,6 +24,19 @@ module Users
         { status: 'Success deleting user', errors: [] }
       else
         { status: nil, errors: object.errors.full_messages }
+      end
+    end
+
+    def login
+      user = User.authenticate params[:email], params[:password]
+
+      if user.blank?
+        { token: nil, errors: ['Invalid email or password.'] }
+      elsif user.archived?
+        { token: nil, errors: ['This User is deactivated.'] }
+      else
+         token = Jwt::TokenGenerator.issue_token user.for_token
+        { token: token, errors: [] }
       end
     end
   end
